Extract NotificationSoundSettings to restore the Swift file length budget
https://github.com/manaflow-ai/cmux/pull/5651 grew
Sources/TerminalNotificationStore.swift (+96 over budget) and
Sources/Feed/FeedCoordinator.swift (+16 over budget) without updating
.github/swift-file-length-budget.tsv, so workflow-guard-tests now fails
on every PR.

Move the self-contained NotificationSoundSettings enum (sound selection,
custom sound staging, Focus/DND suppression, fallback playback, custom
command execution) into its own file. TerminalNotificationStore.swift
drops 2623 -> 1941 lines, well under its 2527 budget, and the budget
ratchets down to the new actual. FeedCoordinator's +16 is inline
authorization-error wiring with no extractable chunk, so its budget is
refreshed to 1255 as known debt already merged on main.

No code changes, pure file move plus pbxproj wiring and budget refresh.
